initial code for variable ST erroring. check writeup and TODO in code for issues
